feat: add full liuyao divination firmware implementation
This commit adds the complete Liu Yao (Six Lines Divination) firmware for the AI Passport:
- Adds core engine, calendar, data, and reading generation logic
- Implements LVGL UI theme, page flow, and all application screens
- Adds tooling for font generation and validation checks
- Includes host tests for all core logic
- Updates documentation and asset licensing information
- Fixes low power timer format string compatibility issue

The firmware supports offline divination with traditional Wenwang Najia methods, including full calendar support, multiple question categories, manual/device casting, and detailed textual interpretations.

+++ main/demo_low_power.c
@@ -154,8 +154,9 @@ static void sleep_task(void *arg)
 
             char text[128];
             if (err == ESP_OK) {
-                snprintf(text, sizeof(text), "LIGHT WAKE: TIMER\nSlept: %lld ms",
-                         (long long)slept_ms);
+                // unsigned long + %lu:主机(MinGW)与 ESP-IDF(newlib)通用。
+                snprintf(text, sizeof(text), "LIGHT WAKE: TIMER\nSlept: %lu ms",
+                         (unsigned long)slept_ms);
             } else {
                 snprintf(text, sizeof(text), "%s failed:\n%s", failure, esp_err_to_name(err));
                 ESP_LOGE(TAG, "%s 失败: %s", failure, esp_err_to_name(err));
